change cycle variables from float to long double
float was leading to overflow, double should be fine
but using long double just to be on the safer side

+++ inc/champsim.h
@@ -75,7 +75,7 @@ extern uint64_t current_core_cycle[NUM_CPUS],
                 last_drc_write_mode,
                 drc_blocks;
 
-extern float master_clock,
+extern long double master_clock,
              next_cpu_event,
              next_dram_event;
